+++ R/monitor_filterDate.R
@@ -60,8 +60,10 @@ monitor_filterDate <- function(
 
   # ----- Validate parameters --------------------------------------------------
 
-  # A little involved to catch the case where the user forgets to pass in 'monitor'
+  MazamaCoreUtils::stopIfNull(startdate)
+  MazamaCoreUtils::stopIfNull(enddate)
 
+  # A little involved to catch the case where the user forgets to pass in 'monitor'
   result <- try({
     if ( !monitor_isValid(monitor) )
       stop("First argument is not a valid 'mts_monitor' object.")
@@ -76,6 +78,20 @@ monitor_filterDate <- function(
 
   if ( monitor_isEmpty(monitor) )
     stop("Parameter 'monitor' has no data.")
+
+  # Deal with missing timezones
+  if ( is.null(timezone) ) {
+    if ( length(unique(monitor$meta$timezone)) == 1 ) {
+      timezone <- monitor$meta$timezone[1]
+    } else {
+      if ( lubridate::is.POSIXct(startdate) ) {
+        timezone <- lubridate::tz(startdate)
+      } else {
+        message("Multiple timezones found and none specified. Using 'UTC'.")
+        timezone <- "UTC"
+      }
+    }
+  }
 
   # ----- Call MazamaTimeSeries function ---------------------------------------